Collapse storage::lock_manager::utils into parent
